requirements.txt
langchain
google-genai
faiss-cpu
pypdf
python-dotenv
streamlit
